Extract upvoteControls function
Added a new module in the utils dir for the upvoteControls function
that's used in the request reducer
